checkpoint

+++ contracts/auction/interfaces/IWETH.sol
@@ -4,12 +4,20 @@ pragma solidity ^0.6.12;
 
 interface IWETH {
     function balanceOf(address account) external view returns (uint256);
+
     function deposit() external payable;
-    function withdraw(uint wad) external; 
-    function totalSupply() external view returns (uint);
-    function approve(address guy, uint wad) external returns (bool);
-    function transfer(address dst, uint wad) external returns (bool);
-    function transferFrom(address src, address dst, uint wad)
-        external
-        returns (bool);
+
+    function withdraw(uint256 wad) external;
+
+    function totalSupply() external view returns (uint256);
+
+    function approve(address guy, uint256 wad) external returns (bool);
+
+    function transfer(address dst, uint256 wad) external returns (bool);
+
+    function transferFrom(
+        address src,
+        address dst,
+        uint256 wad
+    ) external returns (bool);
 }
